Console: move history classes to their own package
