+++ org.metaborg.spoofax.shell.console/src/main/java/org/metaborg/spoofax/shell/client/console/history/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * This package contains implementations (one volatile, one persistent) of
+ * {@link org.metaborg.spoofax.shell.client.IInputHistory}.
+ */
+package org.metaborg.spoofax.shell.client.console.history;
